requirements.txt
matplotlib
pillow
ultralytics